# docs/swagger.yaml
definitions:
  database.Column:
    properties:
      isPKey:
        type: boolean
      name:
        type: string
      required:
        type: boolean
      type:
        type: string
    type: object
  database.Condition:
    properties:
      columns:
        description: предыдущий и текущий столбец таблицы.
        items:
          $ref: '#/definitions/database.QColumn'
        type: array
      operator:
        type: string
    type: object
  database.QColumn:
    properties:
      func:
        description: используется только в select.
        type: string
      name:
        type: string
      payload:
        additionalProperties: {}
        description: специальные данные, привязанные к этому столбцу.
        type: object
      tableKey:
        allOf:
        - $ref: '#/definitions/database.QTableKey'
        description: ключ таблицы, которой принадлежит столбец.
      value:
        description: используется в insert, update, delete и where.
    type: object
  database.QTable:
    properties:
      increment:
        description: приращение имени для создания уникальных псевдонимов.
        type: integer
      name:
        description: имя таблицы.
        type: string
      next:
        description: используется только в select.
        items:
          $ref: '#/definitions/database.QTable'
        type: array
      rule:
        allOf:
        - $ref: '#/definitions/database.Rule'
        description: правило объединения с предыдущей таблицей.
    type: object
  database.QTableKey:
    properties:
      increment:
        description: приращение имени для создания уникальных псевдонимов.
        type: integer
      name:
        description: имя таблицы.
        type: string
    type: object
  database.Rule:
    properties:
      conditions:
        items:
          $ref: '#/definitions/database.Condition'
        type: array
      type:
        type: string
    type: object
  database.Table:
    properties:
      columns:
        items:
          $ref: '#/definitions/database.Column'
        type: array
      name:
        type: string
    type: object
  entity.Query:
    properties:
      columns:
        items:
          $ref: '#/definitions/database.QColumn'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      orderBy:
        description: используется только в select.
        items:
          $ref: '#/definitions/database.QColumn'
        type: array
      sourceId:
        type: string
      table:
        $ref: '#/definitions/database.QTable'
      type:
        type: string
      where:
        items:
          $ref: '#/definitions/database.QColumn'
        type: array
    type: object
  entity.Source:
    properties:
      connected:
        type: boolean
      databaseName:
        type: string
      driver:
        type: string
      host:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
      port:
        type: integer
      username:
        type: string
    type: object
  entity.Widget:
    properties:
      children:
        items:
          $ref: '#/definitions/entity.Widget'
        type: array
      id:
        type: string
      name:
        type: string
      props:
        items:
          type: integer
        type: array
      query:
        description: чтобы ускорить разработку пока так, а дальше посмотрим :/.
        items:
          type: integer
        type: array
      type:
        type: string
    type: object
info:
  contact: {}
paths:
  /queries/execute:
    post:
      parameters:
      - description: запрос
        in: body
        name: query
        required: true
        schema:
          $ref: '#/definitions/entity.Query'
      responses: {}
      tags:
      - запросы
  /sources:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Source'
            type: array
      tags:
      - источники
    patch:
      parameters:
      - description: источник
        in: body
        name: source
        required: true
        schema:
          $ref: '#/definitions/entity.Source'
      responses: {}
      tags:
      - источники
    post:
      parameters:
      - description: источник
        in: body
        name: source
        required: true
        schema:
          $ref: '#/definitions/entity.Source'
      responses: {}
      tags:
      - источники
  /sources/{id}:
    delete:
      parameters:
      - description: идентификатор источника
        in: path
        name: id
        required: true
        type: string
      responses: {}
      tags:
      - источники
    get:
      parameters:
      - description: идентификатор источника
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Source'
      tags:
      - источники
  /sources/{id}/functions:
    get:
      parameters:
      - description: идентификатор источника
        in: path
        name: id
        required: true
        type: string
      responses: {}
      tags:
      - источники
  /sources/{id}/tables:
    get:
      parameters:
      - description: идентификатор источника
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Table'
            type: array
      tags:
      - источники
  /sources/drivers:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      tags:
      - источники
  /widgets:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Widget'
            type: array
      tags:
      - виджеты
    post:
      parameters:
      - description: виджет
        in: body
        name: widget
        required: true
        schema:
          $ref: '#/definitions/entity.Widget'
      responses: {}
      tags:
      - виджеты
  /widgets/{id}:
    delete:
      parameters:
      - description: идентификатор виджета
        in: path
        name: id
        required: true
        type: string
      responses: {}
      tags:
      - виджеты
    get:
      parameters:
      - description: идентификатор виджета
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Widget'
      tags:
      - виджеты
swagger: "2.0"